refactor(chat): Divided ws and api into two controllers
