+++ app/main.ts
@@ -9,11 +9,12 @@ import {
 	InMemoryRepository,
 } from "@infra/in_memory";
 import { KVSRepository } from "@infra/kvs";
+import { apiChatController } from "@presentation/controllers/api_chat_controller";
 import { authController } from "@presentation/controllers/auth_controller";
-import { chatController } from "@presentation/controllers/chat_controller";
 import { pongController } from "@presentation/controllers/pong_controller";
 import { profileController } from "@presentation/controllers/profile_controller";
 import { relationshipController } from "@presentation/controllers/relationship_controller";
+import { webSocketChatController } from "@presentation/controllers/ws_chat_controller";
 import { RegisterUserUsecase } from "@usecase/auth/register_user_usecase";
 import {
 	GetDirectMessagesUsecase,
@@ -95,14 +96,16 @@ const start = async () => {
 		const leaveChatUsecase = new LeaveChatUsecase(chatClientRepository);
 		const getDirectMessagesUsecase = new GetDirectMessagesUsecase(tx);
 		app.register(
-			chatController(
+			webSocketChatController(
 				joinChatUsecase,
 				leaveChatUsecase,
 				sendChatMessageUsecase,
 				sendGameInviteUsecase,
-				getDirectMessagesUsecase,
-				sendDirectMessageUsecase,
 			),
+			{ prefix: "/ws" },
+		);
+		app.register(
+			apiChatController(getDirectMessagesUsecase, sendDirectMessageUsecase),
 			{ prefix: "/api" },
 		);
